Refactor initialState
