+++ src/index.js
@@ -5,30 +5,32 @@ import { Provider } from 'react-redux';
 import { store } from '../src/reducers/books';
 
 //initial state
-// const initialState = {
-//   allBooks: [
-//     { 
-//       bookId: Math.floor((Math.random() * 100) + 1),
-//       title: "The Hermit",
-//       category: "History"
-//     }, 
-//     {
-//       bookId: Math.floor((Math.random() * 100) + 1),
-//       title: "Queen of Endol",
-//       category: "Horror"
-//     },
-//     {
-//       bookId: Math.floor((Math.random() * 100) + 1),
-//       title: "Goals",
-//       category: "Learning"
-//     },
-//     {
-//       bookId: Math.floor((Math.random() * 100) + 1),
-//       title: "Helena of Avalor",
-//       category: "Kids"
-//     }
-//   ]
-// };
+const initialState = {
+  filter: 'George',
+  allBooks: [
+    { 
+      bookId: Math.floor((Math.random() * 100) + 1),
+      title: "The Hermit",
+      category: "History"
+    }, 
+    {
+      bookId: Math.floor((Math.random() * 100) + 1),
+      title: "Queen of Endol",
+      category: "Horror"
+    },
+    {
+      bookId: Math.floor((Math.random() * 100) + 1),
+      title: "Goals",
+      category: "Learning"
+    },
+    {
+      bookId: Math.floor((Math.random() * 100) + 1),
+      title: "Helena of Avalor",
+      category: "Kids"
+    }
+  ]
+};
+
 
 ReactDOM.render(
   <Provider store={store}>
@@ -37,3 +39,4 @@ ReactDOM.render(
   document.getElementById("root")
 );
 
+export default initialState;
