+++ src/reducers/books.js
@@ -1,32 +1,5 @@
-// import { initialState } from "../index";
+import initialState from "../index";
 import { createStore } from 'redux';
-
-const initialState = {
-    filter: 'George',
-    allBooks: [
-      { 
-        bookId: Math.floor((Math.random() * 100) + 1),
-        title: "The Hermit",
-        category: "History"
-      }, 
-      {
-        bookId: Math.floor((Math.random() * 100) + 1),
-        title: "Queen of Endol",
-        category: "Horror"
-      },
-      {
-        bookId: Math.floor((Math.random() * 100) + 1),
-        title: "Goals",
-        category: "Learning"
-      },
-      {
-        bookId: Math.floor((Math.random() * 100) + 1),
-        title: "Helena of Avalor",
-        category: "Kids"
-      }
-    ]
-  };
-  
 
 // booksReducer
